refactor(domain): split entities.rs into identifiers, attributes, and article modules
